# backend/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Internal developer platform API — service catalog, Kubernetes cluster management, and OIDC authentication via Pinniped.",
        "title": "WxOps Portal API",
        "contact": {
            "name": "Platform Team",
            "url": "https://wxops.cloud"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/cli/download/{platform}": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Proxies the latest release asset from the portal Gitea repo using the service-account token.  Requires a valid session cookie.",
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "cli"
                ],
                "summary": "Download wxops CLI binary",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Target platform (linux-amd64, linux-arm64, darwin-amd64, darwin-arm64)",
                        "name": "platform",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "CLI binary",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/auth/callback": {
            "get": {
                "description": "Exchanges the authorization code for tokens, validates the id_token, and writes an encrypted session cookie.",
                "tags": [
                    "auth"
                ],
                "summary": "OIDC callback",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization code",
                        "name": "code",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "CSRF state token",
                        "name": "state",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirect to dashboard",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/auth/me": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Returns sub, username, and group memberships from the session. Available on both /auth/me (unauthenticated, for Next.js SSR) and /api/v1/me (protected).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Current user identity",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.userResponsePayload"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.userResponsePayload": {
            "type": "object",
            "properties": {
                "groups": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "sub": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "CookieAuth": {
            "description": "Encrypted session cookie issued after OIDC login.",
            "type": "apiKey",
            "name": "wxops_session",
            "in": "cookie"
        }
    }
}